requirements.txt
aiohttp
azure-functions